requirements.txt
coincurve
base58
pycryptodome